% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_VennDiagram.r
\name{Create_VennDiagram}
\alias{Create_VennDiagram}
\title{Creates and stores in a current directory a Venn Diagram using 2 gene/element lists.}
\usage{
Create_VennDiagram(GeneList1, GeneList2, varname_1, varname_2)
}
\arguments{
\item{GeneList1}{A list of elements.}

\item{GeneList2}{A list of elements.}

\item{varname_1}{A string containing the variable name of the first variable to be displayed in the plot.}

\item{varname_2}{A string containing the variable name of the second variable to be displayed in the plot.}
}
\value{
A 350x350 png image of a Venn Diagram of the elements in list1 and list2. Image is saved in the current working directory.
}
\description{
\code{Create_VennDiagram} Creates a Venn Diagram using two lists of elements (genes,attributes,cats,pokemons,etc).
}
\examples{
list1 <- c("dog","cat","mouse","ostrich","mole")
list2 <- c("mole","dog","bird","shark")
Create_VennDiagram(list1,list2,"Animals1","Animals2")

}
