% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GTEx_annot_conditional.r
\name{load_GTEX_annot_conditional}
\alias{load_GTEX_annot_conditional}
\title{Load annotation data for MN-AntHorn or Soma-Axon datasets.}
\usage{
load_GTEX_annot_conditional(path, data, data_type = "MN-AH")
}
\arguments{
\item{path}{path and file name of the GTEx expression dataset. rows=gene_ids, cols=samples.}

\item{data}{dataset containing expression data, loaded using the load_GTEX_data_conditional function.}

\item{data_type}{Takes two possible valid options "MN-AH" or "Soma-Axon", to specify which dataset is being loaded, either the Motor Neuron-Anterior horn (GSE18920) or the Soma-Axon (GSE121069) datasets.}
}
\value{
an annotation dataframe containing 3 columns (status: ALS or CTRL, gender: F or M, age: numerical parameter) for the MN-AH, or 2 columns (status: ALS, CTRL, or CTRL_Single; area: Axon, Soma or Single) for the Soma-Axon dataset.
}
\description{
\code{load_GTEX_annot_conditional} Loads MN-AntHorn or Soma-Axon annotation for matching dataset loaded using the load_GTEX_data_conditional function.
}
\examples{
# Loads the annotation dataframe for the soma-axon GSE121069_GEO_rpkms_human.txt dataset.
annot_human_soma_axon <- load_GTEX_annot_conditional("GSE121069_GEO_rpkms_human.txt",data_human_soma_axon,data_type="Soma-Axon")
}
