% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GTEx_data_conditional.r
\name{load_GTEX_data_conditional}
\alias{load_GTEX_data_conditional}
\title{Load GTEx MN-AntHorn or Soma-Axon datasets.}
\usage{
load_GTEX_data_conditional(path, pattern, sep = "\\t")
}
\arguments{
\item{path}{path and file name of the GTEx expression dataset. rows=gene_ids, cols=samples.}

\item{pattern}{grep patter to match samples to be kept for the analysis.}

\item{sep}{sep argument to denote separation used in the expression dataset. Valid options are comma="," tab="\t" space=" " semicolon=";" etc.}
}
\value{
a dataframe containing n rows equal to genes and m columns equal to the samples matching the grep pattern specified.
}
\description{
\code{load_GTEX_data_conditional} Loads MN-AntHorn or Soma-Axon datasets from GTEx in the format rows=gene_ids, cols=samples. Assumes that the first element of the dataset is "NAME".
}
