% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGMA_GSA.r
\name{MAGMA_GSA}
\alias{MAGMA_GSA}
\title{Run MAGMA GSA using gene lists of up and downregulated genes obtained using run_diffExp_analysis.r}
\usage{
MAGMA_GSA(
  tt_filename,
  analysis_type,
  genes.raw_path,
  species = "mouse",
  gene_n
)
}
\arguments{
\item{tt_filename}{diff. Expression file obtained using run_diffExp_analysis function.}

\item{analysis_type}{Accepted values are 'H_MN-AH', 'D_MN-AH', 'H_Soma-Axon', 'D_Soma-Axon', 'Soma_als-ctrl', or 'Axon_als-ctrl'. Same as indicated in run_diffExp_analysis function.}

\item{genes.raw_path}{genes.raw path including filename. genes.raw file is obtained using map_snps_togenes formula.}

\item{species}{Accepted values include "human" or "mouse" default option is mouse.}

\item{gene_n}{Number of genes to be considered up or downregulated. Gene selection is based on t-value from the differential expression analysis.}
}
\value{
Creates Genecovariate files of gene_n up and downregulated genes, and final GSA of both sets of genes in the given GWAS summary statistics.
}
\description{
\code{MAGMA_GSA} Uses a dif. expression analysis file produced using run_diffExp_analysis.r to obtain lists of up and down-regulated genes to be used as gene sets and test for overrepresentation in GWAS summary statistics.
}
\examples{
tt_filename <- run_diffExp_analysis(mouse_D_Soma-Axon_annot,mouse_D_Soma-Axon_data,"/Soma-Axon_Data/","D_Soma-Axon","mouse")
MAGMA_GSA(tt_filename = tt_filename, analysis_type = "D_Soma-Axon", genes.raw_path = "/magmafiles/als.genes.raw", species = "mouse", gene_n = 250)

}
