% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_snps_togenes.r
\name{map_snps_togenes}
\alias{map_snps_togenes}
\title{Write .genes.annot and .genes.raw files required for GSA analysis using MAGMA.}
\usage{
map_snps_togenes(
  gwas_path,
  N = NULL,
  genloc_filepath,
  genome_ref_path,
  analysis_type,
  species
)
}
\arguments{
\item{gwas_path}{GWAS summary statistics path including file name.}

\item{N}{N parameter of GWAS summary statistics can be set to the total N. If no N column is present in GWAS summary statistics, this parameter has to be specified by the user.}

\item{genloc_filepath}{Human gene location file path.}

\item{genome_ref_path}{Human genome reference path and prefix of .bed, .bim, .fam files. If .synonyms file with the same prefix is present in the same folder it is automatically detected.}

\item{analysis_type}{Accepted values are 'H_MN-AH', 'D_MN-AH', 'H_Soma-Axon', 'D_Soma-Axon', 'Soma_als-ctrl', or 'Axon_als-ctrl'. Same as indicated in run_diffExp_analysis function.}

\item{species}{Gene expression species. Accepted values include "human" or "mouse" default option is mouse.}
}
\value{
Path of the .genes.raw file to be used for MAGMA_GSA. Two files (.genes.annot and .genes.raw) are generated and stored. Maps snps in GWAS summary statistics into genes and produces a gene-level p-value to be used for the GSA.
}
\description{
\code{map_snps_togenes} This function precedes MAGMA_GSA function if now .genes.raw file is present. It is used to prepare GWAS file for GSA.
}
\examples{
genes.raw_path <- map_snps_togenes(gwas_path = "/Sumstats_ALS/ALS_sumstats_EUR_only.txt",N = NULL,genloc_filepath = "/genloc_files/NCBI37.3.gene.loc", genome_ref_path = "/g1000/g1000_eur", analysis_type = "D_Soma-Axon", species = "mouse")

}
