% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.tt.r
\name{prep.tt}
\alias{prep.tt}
\title{Prepare variables and dataset for diff. expression analysis. This function is used by run_als_diffExp_analysis.}
\usage{
prep.tt(exp_data, mod, coef, species, analysis_type, file_name)
}
\arguments{
\item{exp_data}{dataframe containing expression data where rows=gene_ids and cols=sample_ids.}

\item{mod}{Design or model matrix for Diff. Exp. Analysis.}

\item{coef}{parameter taken by topTable function to indicate which variable of interest is to be extracted from the linear model.}

\item{species}{species from which expression data was derived. Valid options are either "mouse" or "human".}

\item{analysis_type}{string indicating analysis to be carried out. Allowed strings are 'H_MN-AH', 'D_MN-AH', 'H_Soma-Axon', 'D_Soma-Axon', 'Soma_als-ctrl', or 'Axon_als-ctrl'.}

\item{file_name}{Diff. Expression Analysis file name.}
}
\value{
a file is stored in the subfolder Results/Tables of the working directory and a dataframe with the restuls of the diff. expression analysis is returned.
}
\description{
\code{prep.tt} Function used by run_diffExp_analysis function to prepare data and variables for diff expression analysis of MN-AH or Soma-Axon data.
}
