% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkm_to_tpm.r
\name{rpkm_to_tpm}
\alias{rpkm_to_tpm}
\title{Convert an expression dataset in rpkm to tpm.}
\usage{
rpkm_to_tpm(dataset)
}
\arguments{
\item{dataset}{expressed in rpkm. rows = genes and columns = samples.}
}
\value{
a dataset expressed in tpm.
}
\description{
\code{rpkm_to_tpm} convert a dataset expressed in rpkm into tpm.
}
