% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diffExp_analysis.r
\name{run_diffExp_analysis}
\alias{run_diffExp_analysis}
\title{Run differential expression analysis for GTEx data previously loaded using load_GTEx_data_conditional.}
\usage{
run_diffExp_analysis(annot, data, expr_path, analysis_type, species = "human")
}
\arguments{
\item{annot}{dataframe containing the annotation dataframe of the expr. data, created using load_GTEx_annot_conditional function.}

\item{data}{dataframe containing expression data, loaded using the load_GTEX_data_conditional function.}

\item{expr_path}{Path to the file containing the expression data. Path doesn't include the file name.}

\item{analysis_type}{string indicating analysis to be carried out. Permitted strings are 'H_MN-AH', 'D_MN-AH', 'H_Soma-Axon', 'D_Soma-Axon', 'Soma_als-ctrl', or 'Axon_als-ctrl'.}

\item{species}{Permitted strings are either "mouse" or "human".}
}
\value{
path to the differential expression file.
}
\description{
\code{run_diffExp_analysis} Runs differential expression analysis for GTEx MN-AH or Soma-Axon data.
}
\examples{
# Runs differential expression analysis for mouse Soma-Axon data.
tt_mouse_SomaAxon <- run_diffExp_analysis(annot = annot_mouse_axon_soma, data = mouse_axon_soma, expr_path = "/Users/AnaCrisGlez/Soma_Axon_RNA-Seq/", analysis_type = "D_Soma-Axon", species = "mouse")

}
