% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gene_covarfile.r
\name{write_gene_covarfile}
\alias{write_gene_covarfile}
\title{Write Gene covariate file to use as Gene Set for GSA. This function is a dependency of MAGMA_GSA.r}
\usage{
write_gene_covarfile(geneset, varname, filename)
}
\arguments{
\item{geneset}{geneset a list of up or downregulated genes, obtained using MAGMA_GSA.r}

\item{varname}{varname Geneset name, generated by MAGMA_GSA.r: UR/DR_analysistype_species.}

\item{filename}{Genecovar filename, generated by MAGMA_GSA.r: GeneCovar_DR/UR_analysistype_species.}
}
\value{
Genecovar file containing the entrez ids of the n up or downregulated genes obtained using MAGMA_GSA function
}
\description{
\code{write_gene_covarfile} Writes a covariate file containing entrez gene ids of up or downregulated genes.
}
